requirements.txt
google-genai>=1.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
requests>=2.31.0
pyyaml>=6.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
Pillow>=10.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
supabase>=2.0.0
pydantic>=2.0.0

# Test dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
